started decomposing main function in index (splitting into modules)
