+++ src/parser/helpers/common.js
@@ -1,7 +1,7 @@
 'use strict';
 
 function arrayFromLength(number) {
-  return Array.from(new Array(number).keys()).map(k => k+1)
+  return Array.from(new Array(number).keys()).map(k => k + 1);
 }
 
 module.exports = {
